requirements.txt
fastapi==0.115.6
uvicorn==0.30.6

pymongo==4.8.0
python-dotenv==1.0.1

openai==2.15.0
python-multipart

Pillow